backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Snowflake (with pandas support)
snowflake-snowpark-python[pandas]==1.11.1
snowflake-connector-python[pandas]==3.6.0
cryptography>=41.0.0

# Database and data processing
pandas>=2.0.0,<2.2.0
numpy>=1.24.0,<2.0.0
pyarrow>=12.0.0

# analytics
scikit-learn>=1.3.0
umap-learn>=0.5.4
matplotlib>=3.7.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
colorama==0.4.6
pyperclip==1.8.2
tabulate==0.9.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.14.6

sentence-transformers
